+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+
+### Fixed
+- Fixed an error that occurred when executing a GraphQL query if a Width field existed in the Craft install
+
 ## 4.1.0 - 2022-10-20
 
 ### Added